requirements.txt
boto3>=1.42,<1.43
requests>=2.32.3,<2.33
python-dotenv>=1.0
protobuf>=4.25,<6
gtfs-realtime-bindings>=2.0
pandas>=2.0
snowflake-connector-python[pandas]>=3.0
pyspark>=3.4,<3.5
cryptography>=42
